+++ modules/navigation_3d/editor/navigation_link_3d_gizmo_plugin.h
